Agregado de SweetAlert de carga a la mayoria de vistas

+++ src/public/js/cart.js
@@ -7,21 +7,30 @@ btnDelete.forEach(btn=>{
     btn.addEventListener('click', evt=>{
         evt.stopPropagation();
         const product_id = btn.value;
-        fetch(`/api/carts/${cart_id.value}/product/${product_id}`, {method: 'DELETE'})
-            .then(res=>res.json())
-            .then(data=>{
-                console.log(data);
-                if(data.status === 'success'){
-                    Swal.fire({
-                        text: `Producto eliminado del carrito.`,
-                        icon: 'success',
-                        timer: 1500,
-                        showConfirmButton: false
-                    }).then(()=>{
-                        location.reload();
+
+        Swal.fire({
+            title: 'Borrando producto del carrito',
+            html: 'Aguarde unos momentos...',
+            didOpen: () => {
+                Swal.showLoading()
+    
+                fetch(`/api/carts/${cart_id.value}/product/${product_id}`, {method: 'DELETE'})
+                    .then(res=>res.json())
+                    .then(data=>{
+                        console.log(data);
+                        if(data.status === 'success'){
+                            Swal.fire({
+                                text: `Producto eliminado del carrito.`,
+                                icon: 'success',
+                                timer: 1500,
+                                showConfirmButton: false
+                            }).then(()=>{
+                                location.reload();
+                            })
+                        }
                     })
-                }
-            })
+            }
+        })
     })
 })
 
@@ -44,27 +53,34 @@ btnEndCompra.addEventListener('click', evt=>{
             url = `/api/carts/${cart_id.value}/purchase`;
         }
 
-        console.log(url);
-        fetch(url)
-            .then(res=>res.json())
-            .then(data=>{
-                if(data.status == 'success'){
-                    let html = `Compra finalizada! Su numero de ticket es: ${data.payload.ticket}.`;
-                    if(data.payload.products_out_of_stock.length > 0){
-                        html += `<br>Los siguientes articulos no se compraron por falta de stock<br><ul>`;
-                        data.payload.products_out_of_stock.forEach(prod=>{
-                            html += `<li>${prod.title} - cant: ${prod.quantity} - stock: ${prod.stock}</li>`;
-                        })
-                        html += `</ul>`;
-                    }
-
-                    Swal.fire({
-                        html,
-                        icon: 'success',
-                    }).then(()=>{
-                        location.reload();
+        Swal.fire({
+            title: 'Realizando la compra',
+            html: 'Aguarde unos momentos...',
+            didOpen: () => {
+                Swal.showLoading()
+    
+                fetch(url)
+                    .then(res=>res.json())
+                    .then(data=>{
+                        if(data.status == 'success'){
+                            let html = `Compra finalizada! Su numero de ticket es: ${data.payload.ticket}.`;
+                            if(data.payload.products_out_of_stock.length > 0){
+                                html += `<br>Los siguientes articulos no se compraron por falta de stock<br><ul>`;
+                                data.payload.products_out_of_stock.forEach(prod=>{
+                                    html += `<li>${prod.title} - cant: ${prod.quantity} - stock: ${prod.stock}</li>`;
+                                })
+                                html += `</ul>`;
+                            }
+        
+                            Swal.fire({
+                                html,
+                                icon: 'success',
+                            }).then(()=>{
+                                location.reload();
+                            })
+                        }
                     })
-                }
-            })
+            }
+        })
     })
 })
